refactor: major implementation refactor
Remove the legacy mechanism over to a new mechansim
based on LoanEvents and day actions.

The basis of day actions computation should NOT be influenced by LoanEvents,
rather, the LoanEvents should reprocess the day actions.

Currently, only the initialization LoanEvent is implemented.
We are missing the LoanScheduleExtra to be on feature parity with legacy implementation.